[P4ADEV-2382] add check iuv
